// src/interfaces.ts
export interface SearchData {
  image: string;
  janCode: string;
  name: string;
  price: number;
  seller: string;
  shipping: string;
  url: string;
}

export interface RegisteredData {
  image: string;
  item_code: string;
  name: string;
  price: number;
  seller: string;
  shipping: string;
  url: string;
  border_price: number;
}